Proof of concept for a sorting algorithm includes a sudo-sort
The idea here was to build a O(n) sorting algorithm. The resulting algorithm is O(n + r) where n is the number of numbers and r is the range of numbers. NOTE: this is not fully fleshed out and needs additional testing for adding in numbers and the subsequent ratification.